Restructure tools
